requirements.txt
async-timeout
wavelink==3.4.1
wavelink[spotify]
discord.py>=2.0.0
python-dotenv>=0.19.0
aiofiles>=0.8.0